reorganize tui tests
