+++ tests/tui/__init__.py
@@ -1,0 +1,3 @@
+"""Tests for the ruff-sync TUI."""
+
+from __future__ import annotations
